wip update
Pressing P will print out a photo, theres still a lot of work left to be done on it though including fixing the lighting issue with it as well as saving it locally to a folder. theres a good tutorial that has 4 parts and its like an hour total and its a fair bit of work to get everything running so this might come later

Camera will now cause player to become invisible when the camera angle is in-between (20 and 45 degrees) --- right now its only off and on, with clamping the visibility to make the player material bot translucent is a lot harder since it would require new material graphs. So now you can look up basically and see shit without seeing the back of the bot's head.