Maintenance and minor improvements to gameplay and UI
The lights puzzles in area 2 were not working so i went ahead and redid some of the logic for how they are processed and it seems to be more future proof now and functions normally. I went ahead and added a selection process to buying items -- its similar to the soul code widget menu where you can submit an option or leave. Also changed some platforms to be larger in area 2 that move around since some playtesting revealed it was still too difficult for ppl at times ;-;

Lastly, 2 The bouncing back and forth glitch with the playerball happens in the elevator up and down puzzle. theres a bug here i wasn't able to reproduce... it MIGHT be because the build was tested on a low power laptop with all the settings turned down to low? framerate may affected it?
In any case if it is framerate like i suspect at this time that may be grounds for redoing the entire ball transfer system with the playerball tbh.